SIG-1503 further refactor
