+++ api/app/signals/apps/sigmax/stuf_protocol/exceptions.py
@@ -1,0 +1,2 @@
+class SigmaxException(Exception):
+    pass
